+++ app/src/app/seed-module/seed.config.js
@@ -9,22 +9,29 @@
     function moduleConfig($stateProvider, triMenuProvider) {
 
         $stateProvider
-        .state('triangular.admin-default.seed-page', {
+        .state('triangular.seed-page', {
             url: '/seed-module/seed-page',
             templateUrl: 'app/seed-module/seed-page.tmpl.html',
             // set the controller to load for this page
             controller: 'SeedPageController',
-            controllerAs: 'vm'
+            controllerAs: 'vm',
+            // layout-column class added to make footer move to
+            // bottom of the page on short pages
+            data: {
+                layout: {
+                    contentClass: 'layout-column'
+                }
+            }
         });
 
         triMenuProvider.addMenu({
             name: 'Seed Module',
-            icon: 'zmdi zmdi-grade',
+            icon: 'fa fa-tree',
             type: 'dropdown',
             priority: 1.1,
             children: [{
                 name: 'Start Page',
-                state: 'triangular.admin-default.seed-page',
+                state: 'triangular.seed-page',
                 type: 'link'
             }]
         });